+++ app/views.py
@@ -1,7 +1,6 @@
 import json
 import random
 import requests
-from typing import Optional, cast
 from datetime import datetime, timedelta
 
 from django.contrib import auth
@@ -57,97 +56,6 @@ from app.academic_utils import (
     get_search_results,
 )
 
-
-
-class IndexView(SecureTemplateView):
-
-    login_required = False
-    template_name = 'index.html'
-
-    def dispatch_prepare(self, method: str) -> SecureView.HandlerType:
-        match method:
-            case 'get':
-                return (self.user_get
-                        if self.request.user.is_authenticated else
-                        self.visitor_get)
-            case 'post':
-                return self.prepare_login()
-            case _:
-                return self.default_prepare(method)
-
-    def visitor_get(self) -> HttpResponse:
-        # Modify password
-        # Seems that after modification, log out by default?
-        if self.request.GET.get('modinfo') is not None:
-            succeed("修改密码成功!", self.extra_context)
-        return self.render()
-
-    def user_get(self) -> HttpResponse:
-        self.request = cast(UserRequest, self.request)
-        # Special user
-        self.valid_user_check(self.request.user)
-
-        # Logout
-        if self.request.GET.get('is_logout') is not None:
-            auth.logout(self.request)
-            return self.render()
-
-        return self.redirect('welcome')
-
-    def valid_user_check(self, user: User):
-        # Special user
-        if not user.is_valid():
-            self.permission_denied(
-                f'“{user.get_full_name()}”不存在成长档案，您可以登录其他账号'
-            )
-
-    def prepare_login(self) -> SecureView.HandlerType:
-        assert 'username' in self.request.POST
-        assert 'password' in self.request.POST
-        _user = self.request.user
-        assert not _user.is_authenticated or not cast(User, _user).is_valid()
-        username = self.request.POST['username']
-        # Check weather username exists
-        if not User.objects.filter(username=username).exists():
-            # Allow org to login with orgname
-            org = Organization.objects.filter(oname=username).first()
-            if org is None:
-                return self.wrong('用户名不存在')
-            username = org.get_user().username
-        self.username = username
-        self.password = self.request.POST['password']
-        return self.login
-
-    def login(self) -> HttpResponse:
-        # Try login
-        userinfo = auth.authenticate(username=self.username, password=self.password)
-        if userinfo is None:
-            return self.wrong('密码错误')
-
-        # special user
-        auth.login(self.request, userinfo)
-        self.request = cast(UserRequest, self.request)
-        self.valid_user_check(self.request.user)
-
-        # first time login
-        if self.request.user.first_time_login:
-            return self.redirect('modpw')
-
-        # Related account
-        # When login as np, related org accout is also available
-        update_related_account_in_session(self.request, self.username)
-
-        # If origin is present and valid, redirect
-        # Otherwise, redirect to welcome page
-        origin = self.request.GET.get('origin')
-        if origin and self._is_origin_safe(self.request, origin):
-            return self.redirect(origin)
-        else:
-            return self.redirect('welcome')
-
-    def _is_origin_safe(self, request: HttpRequest,
-                         origin: Optional[str] = None) -> bool:
-        return origin is None or origin.startswith('/')
 
 
 @login_required(redirect_field_name="origin")
@@ -1494,12 +1402,6 @@ def authRegister(request: HttpRequest):
 
 
 @logger.secure_view()
-def logout(request: HttpRequest):
-    auth.logout(request)
-    return HttpResponseRedirect("/index/")
-
-
-@logger.secure_view()
 def get_stu_img(request: HttpRequest):
     stuId = request.GET.get("stuId")
     if stuId is not None:
@@ -2079,3 +1981,4 @@ def notifications(request: HttpRequest):
     bar_display = utils.get_sidebar_and_navbar(request.user,
                                                navbar_name="通知信箱")
     return render(request, "notifications.html", locals())
+
